-modifie int to long id

+++ VirtualShelter/src/main/java/hu/bme/aut/vshelter/dal/UserFacade.java
@@ -34,7 +34,7 @@ public interface UserFacade {
 	 * @return
 	 * @throws VirtualShelterException
 	 */
-	public int getUserIdfromEmail(String email) throws VirtualShelterException;
+	public long getUserIdfromEmail(String email) throws VirtualShelterException;
 	
 
 }
